Change typ to avoid warning.
git-svn-id: http://svn.research.cc.gatech.edu/kaos/evpath/trunk@11010 8ba6a8b9-845b-0410-94f0-974be2f5df47

+++ tests/thin_client.c
@@ -53,7 +53,7 @@ main(int argc, char **argv)
     FMFormat ioformat;
     FMContext fmc;
     simple_rec rec;
-    int conn;
+    long conn;
     char *comment = strdup("Stone xxxxx");
 
     if ((argc != 4) || (sscanf(argv[2], "%d", &remote_port) != 1)
